apps/convert/requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
numpy==2.0.0
pandas==2.2.2
protobuf==5.27.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
